chore: extends button to support disabled state, support of links,
