+++ components/button/button.interface.ts
@@ -2,10 +2,12 @@ import { EButtonIconPosition, EButtonSizes, EButtonTypes } from './button.enum';
 import { EIConTypes } from '../icon/icon.enum';
 
 export interface IButtonProps {
-  type: EButtonTypes;
+  type?: EButtonTypes;
   size?: EButtonSizes;
   label: string;
   onClickEvent: () => void;
   icon?: EIConTypes;
   iconPosition?: EButtonIconPosition;
+  href?: string;
+  disabled?: boolean;
 }
